refactored the add to mappings function. Moved most of the logic to the channel object. Fixed function occluded warnings in Shimmer and twinkle functions. fixed consolidate mappings functionality to work on an arbitrary list of source channels bound to an arbitrary destination element.

+++ Modules/Sequence/LightOrama/LightOramaAutoPopulateElementsForm.cs
@@ -26,7 +26,7 @@ namespace VixenModules.SequenceType.LightOrama
 		/// </summary>
 		List<LorChannelMapping> m_mappings = new List<LorChannelMapping>();
 
-		private UInt64 m_mappingCount = 0;
+		private int m_mappingCount = 0;
 
 		public LightOramaAutoPopulateElementsForm(LightOramaSequenceData parsedLorSequence, List<LorChannelMapping> mappings)
 		{
